refactor(media): extract DeploymentLinkPill into shared module
